utils/imdb2sql/requirements.txt
IMDbPY==6.8
